added leaderboard class method to game model and switched game_questions controller to questions controller
