--- app/helpers/game_questions_helper.rb
@@ -1,2 +1,0 @@
-module GameQuestionsHelper
-end
