--- test/controllers/game_questions_controller_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class GameQuestionsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
